requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
sqlmodel==0.0.18
mysqlclient==2.2.4 # MySQL 드라이버
python-dotenv==1.0.1 # .env 파일 로드를 위해 추가